+++ audio_synch_tool/plotters.py
@@ -31,15 +31,37 @@ __author__ = "Andres FR"
 # #############################################################################
 
 
+def sample_to_timestamp_formatter(val, pos, samplerate):
+    """
+    This function can be passed to ``plt.FuncFormatter`` to generate custom
+    tick labels. It fulfills the interface (val, pos) -> str.
+
+    Specifically, converts ``val`` number representing a sample into a string
+    showing the corresponding elapsed time since sample 0, asuming the given
+    samplerate.
+
+    :param number val: the axis value where the tick goes
+    :param int pos: from 0 (left margin) to num_ticks+2 (right
+      margin)
+    :param number samplerate: A number so that seconds = val / samplerate
+    :returns: A string in the form "{X days} h:m:s.ms"
+
+    Usage example::
+      ax.xaxis.set_major_formatter(plt.FuncFormatter(fn))
+    """
+    ts = Timestamp(val, samplerate)
+    ts_str = str(ts)[:-3]  # keep up to miliseconds only
+    return ts_str
+
 # #############################################################################
 # ## PLOTTER CLASSES
 # #############################################################################
 
 
-class AudioPlotter(object):
+class DownsampledPlotter1D(object):
     """
     This class generates matplotlib figures that plot 1-dimensional
-    arrays. Since audio arrays can be quite long, it features a built-in
+    arrays. Since some arrays can be quite long, it features a built-in
     downsampling mechanism that plots a (configurable) number of points
     at most. The downsampling mechanism is based on this code:
 
@@ -73,21 +95,23 @@ class AudioPlotter(object):
         self.max_datapoints = max_datapoints
         self.samplerate = samplerate
 
-    def _downsample_arr(self, xstart, xend):
+    @staticmethod
+    def _downsample_arr(xstart, xend, x_vals, y_vals, max_datapoints):
         """
         Reference:
         https://matplotlib.org/3.1.0/gallery/event_handling/resample.html
         """
         # get the points in the view range
-        mask = (self.arange > xstart) & (self.arange < xend)
+        mask = (x_vals > xstart) & (x_vals < xend)
         # dilate the mask by one to catch the points just outside
         # of the view range to not truncate the line
         mask = np.convolve([1, 1], mask, mode='same').astype(bool)
         # sort out how many points to drop
-        ratio = int(max(np.sum(mask) // self.max_datapoints, 1))
+        ratio = int(max(np.sum(mask) // max_datapoints, 1))
+        print(">>>>>>><", ratio)
         # mask data
-        xdata = self.arange[mask]
-        ydata = self.arr[mask]
+        xdata = x_vals[mask]
+        ydata = y_vals[mask]
         # downsample data
         xdata = xdata[::ratio]
         ydata = ydata[::ratio]
@@ -95,7 +119,8 @@ class AudioPlotter(object):
             len(ydata), np.sum(mask)))
         return xdata, ydata
 
-    def _update_ax(self, ax):
+    @classmethod
+    def _update_ax(cls, ax, ax_line, x_vals, y_vals, max_datapoints):
         """
         Reference:
         https://matplotlib.org/3.1.0/gallery/event_handling/resample.html
@@ -103,8 +128,10 @@ class AudioPlotter(object):
         # Update the line
         lims = ax.viewLim
         xstart, xend = lims.intervalx
-        self.axx.set_data(*self._downsample_arr(xstart, xend))
+        ax_line.set_data(*cls._downsample_arr(
+            xstart, xend, x_vals, y_vals, max_datapoints))
         ax.figure.canvas.draw_idle()
+
 
     def make_fig(self, num_xticks=10, num_yticks=10, tick_fontsize=7):
         """
@@ -117,117 +144,240 @@ class AudioPlotter(object):
         https://matplotlib.org/3.1.0/gallery/event_handling/resample.html
         """
         fig, ax = plt.subplots()
-        self.axx, = ax.plot(self.arange, self.arr, '-')
+        ax_line, = ax.plot(self.arange, self.arr, '-')
         # set number of ticks
         ax.locator_params(axis="x", nbins=num_xticks)
         ax.locator_params(axis="y", nbins=num_yticks)
         #
+        ax.xaxis.tick_top()
         plt.setp(ax.xaxis.get_majorticklabels(), rotation=30,
-                 fontsize=tick_fontsize, family="DejaVu Sans", ha="right")
+                 fontsize=tick_fontsize, family="DejaVu Sans", ha="left")
         plt.setp(ax.yaxis.get_majorticklabels(),
                  fontsize=tick_fontsize, family="DejaVu Sans")
-
         ax.set_autoscale_on(False)  # Otherwise, infinite loop
-        ax.callbacks.connect('xlim_changed', self._update_ax)
+        ax.callbacks.connect('xlim_changed', lambda axx: self._update_ax(
+            axx, ax_line, self.arange, self.arr, self.max_datapoints))
         ax.set_xlim(0, len(self.arr))
         ax.xaxis.grid(True)
 
-
         if self.samplerate is not None:
-            def formatter(val, pos):
-                """
-                This function can be passed to ``plt.FuncFormatter``
-                to generate custom tick labels. It fulfills the interface
-                (val, pos) -> str.
-
-                Specifically, given a ``val`` representing a sample,
-                returns a string showing the corresponding elapsed time
-                since sample 0, asuming ``self.samplerate``.
-
-                :param number val: the axis value where the tick goes
-                :param int pos: from 0 (left margin) to num_ticks+2 (right
-                  margin)
-                :returns: A string in the form "{X days} h:m:s.ms"
-
-                Usage example::
-                  ax.xaxis.set_major_formatter(plt.FuncFormatter(fn))
-                """
-                ts = Timestamp(val, self.samplerate)
-                ts_str = str(ts)[:-3]  # keep up to miliseconds only
-                return ts_str
-            ax.xaxis.set_major_formatter(plt.FuncFormatter(formatter))
+            f = plt.FuncFormatter(lambda val, pos:
+                                  sample_to_timestamp_formatter(
+                                      val, pos, self.samplerate))
+            ax.xaxis.set_major_formatter(f)
         return fig
 
 
-
-
-class MultiTrackPlotter(object):
+class MultipleDownsampledPlotter1D(DownsampledPlotter1D):
     """
     """
-    # plt.rcParams["patch.force_edgecolor"] = True
     FIG_ASPECT_RATIO = (10, 8)
-    # FIG_WIDTH_RATIOS = [5, 5]
-    # # FIG_HEIGHT_RATIOS = [10,3,10] # line, margin, line
-    # FIG_MARGINS = {"top":0.8, "bottom":0.2, "left":0.1, "right":0.9}
     FIG_MARGINS = {"top": 0.8, "bottom": 0.05, "left": 0.05, "right": 0.95,
                    "hspace": 0.5, "wspace": 0.5}
-    # RATIO_RANGE_AXIS = 1.15 # FOR 1.0, THE AXIS WILL COVER EXACTLY UNTIL THE MAXIMUM
-    # RATIO_LABELS_DISTANCE = 0.02 # 0.05 MEANS THE DISTANCE BETWEEN BARS AND LABELS IS 5% OF THE AXIS
-    # #
-    FIG_TITLE_FONTSIZE = 15
-    # AX_TITLE_FONTSIZE = 18
-    # # TABLE_FONTSIZE = 9
-    # TICKS_FONTSIZE = 15
-    # CONFMAT_NUMBER_COLOR = "cyan"
-    # CONFMAT_NUMBER_FONTSIZE = 12
-    # NUM_YTICKS = 13
-    # LABELS_FONTSIZE = 12
-    # LEGEND_FONTSIZE = 15
-    # BAR_WIDTH = 0.85
-    # FPOS_COLOR = "#b0b000"
-    # FNEG_COLOR = "#e6194b"
 
 
-    def __init__(self):
+    def __init__(self, arrays, samplerates=None, max_datapoints=10000,
+                 shared_plots=None):
         """
+        :param list arrays: a list of N 1-dimensional arrays
+        :param list samplerates: If not None, a list of N positive numbers (or
+          ``None`` entries), each one corresponding to the nth given array.
+          If an entry is ``None``, the x-axis will show sample index. Otherwise
+          the sample index will be converted to a timestamp. If the whole
+          argument is ``None``, all axes will show sample index.
+        :param int max_datapoints: Expected to be a positive integer,
+          the number of plotted datapoints that each plot will (approximately)
+          have.
+        :param list shared_plots: If not None, a list of N boolean values, each
+          one corresponding to a given array. All arrays with True value will
+          share the same x axis, i.e. when the user scrolls or zooms across
+          one of them, the rest will change the same way. If None, all will be
+          treated as False, i.e. no x axis will be shared.
         """
-        pass
+        #
+        self.N = len(arrays)
+        assert self.N >= 1, "empty array list?"
+        for arr in arrays:
+            assert len(arr.shape) == 1, "1D array expected!"
+        assert max_datapoints > 0, "positive max_datapoints expected!"
+        if samplerates is not None:
+            assert len(samplerates) == len(arrays),\
+                "Number of samplerates must equal number of arrays!"
+            for sr in samplerates:
+                if sr is not None:
+                    assert sr > 0, "all samplerates must be positive!"
+        if shared_plots is not None:
+            assert len(shared_plots) == len(arrays),\
+                "Number of shared_plots must equal number of arrays!"
+        #
 
-    def make_fig(self, fig_title, audio_arr, total, x_shared):
+        self.arrays = arrays
+        self.aranges = [np.arange(len(arr)) for arr in arrays]
+        self.max_datapoints = max_datapoints
+        self.samplerates = ([None for _ in range(self.N)]
+                            if samplerates is None else samplerates)
+        self.shared_plots = ([False for _ in range(self.N)]
+                            if shared_plots is None else shared_plots)
+
+        self.shared_idxs = [idx for idx, b in enumerate(self.shared_plots)
+                            if b]
+        self.non_shared_idxs = [i for i in range(self.N)
+                                if i not in self.shared_idxs]
+
+
+    def make_fig(self, num_xticks=10, num_yticks=10, tick_fontsize=7,
+                 tick_rot_deg=30):
         """
+        :returns: A matplotlib Figure containing the array given at
+          construction. The interactive plot of the figure will react
+          to the user's zooming by showing approximately
+          ``self.max_datapoints`` number of samples.
+
+        Reference:
+        https://matplotlib.org/3.1.0/gallery/event_handling/resample.html
         """
         # define and configure plt figure
         fig = plt.figure(figsize=self.FIG_ASPECT_RATIO)
-        gs = list(gridspec.GridSpec(total, 1))
-        axes = [plt.subplot(gs[0])]
-        # add shared-x axes
-        axes += [plt.subplot(gsi, sharex=axes[0])
-                 for gsi in gs[1: x_shared]]
-        # add remaining axes:
-        axes += [plt.subplot(gsi) for gsi in gs[x_shared:]]
+        gs = list(gridspec.GridSpec(self.N, 1))
+        # first create non-shared axes
+        axes = [plt.subplot(gs[i]) for i in self.non_shared_idxs]
 
+        # then create the first shared, the rest with share with the first
+        if self.shared_idxs:
+            shared_axes = [plt.subplot(gs[self.shared_idxs[0]])]
+            for idx in self.shared_idxs[1:]:
+                shared_axes.insert(idx, plt.subplot(gs[idx],
+                                                    sharex=shared_axes[0]))
+            for idx, ax in zip(self.shared_idxs, shared_axes):
+                axes.insert(idx, ax)
+
+        # plots
+        lines = [ax.plot(rng, arr, "-")[0]
+                 for arr, rng, ax in zip(self.arrays, self.aranges, axes)]
+
+        # configure axes ticks and labels
+        for ax, line, arr, rng, sr in zip(axes, lines, self.arrays,
+                                          self.aranges, self.samplerates):
+            # set vertical grid
+            ax.xaxis.grid(True)
+
+            # number of x and y ticks
+            ax.locator_params(axis="x", nbins=num_xticks)
+            ax.locator_params(axis="y", nbins=num_yticks)
+
+            # alignment, font and rotation
+            plt.setp(ax.xaxis.get_majorticklabels(), rotation=tick_rot_deg,
+                     fontsize=tick_fontsize, family="DejaVu Sans", ha="right")
+            plt.setp(ax.yaxis.get_majorticklabels(),
+                     fontsize=tick_fontsize, family="DejaVu Sans")
+
+            # axes downsampling
+            ax.callbacks.connect('xlim_changed', lambda _: self._update_ax(
+                ax, line, rng, arr, self.max_datapoints))
+            ax.set_xlim(0, len(arr))
+
+            # finally adapt labels to given sample rates
+            if sr is not None:
+
+                f = plt.FuncFormatter(lambda val, pos:
+                                      sample_to_timestamp_formatter(
+                                          val, pos, sr))
+                print(">>>>>>>>>>", f.func(100000, None))
+                ax.xaxis.set_major_formatter(f)
         #
-        info = "\nNo. axes=%d, dep. axes=%d" % (total, x_shared)
-        fig.suptitle(fig_title + info, fontsize=self.FIG_TITLE_FONTSIZE,
-                     fontweight="bold")
-        fig.canvas.set_window_title(fig_title)
-
-        axes[0].plot(audio_arr)
-
-        for ax in axes[1:]:
-            ax.plot(range(10), range(10))
         fig.subplots_adjust(**self.FIG_MARGINS)
         return fig
 
 
+# class AudioAndMvnPlotter(DownsampledPlotter1D):
+#     """
+#     """
+#     # plt.rcParams["patch.force_edgecolor"] = True
+#     FIG_ASPECT_RATIO = (10, 8)
+#     # FIG_WIDTH_RATIOS = [5, 5]
+#     # # FIG_HEIGHT_RATIOS = [10,3,10] # line, margin, line
+#     # FIG_MARGINS = {"top":0.8, "bottom":0.2, "left":0.1, "right":0.9}
+#     FIG_MARGINS = {"top": 0.8, "bottom": 0.05, "left": 0.05, "right": 0.95,
+#                    "hspace": 0.5, "wspace": 0.5}
+#     # RATIO_RANGE_AXIS = 1.15 # FOR 1.0, THE AXIS WILL COVER EXACTLY UNTIL THE MAXIMUM
+#     # RATIO_LABELS_DISTANCE = 0.02 # 0.05 MEANS THE DISTANCE BETWEEN BARS AND LABELS IS 5% OF THE AXIS
+#     # #
+#     FIG_TITLE_FONTSIZE = 15
+#     # AX_TITLE_FONTSIZE = 18
+#     # # TABLE_FONTSIZE = 9
+#     TICKS_FONTSIZE = 7
+#     # CONFMAT_NUMBER_COLOR = "cyan"
+#     # CONFMAT_NUMBER_FONTSIZE = 12
+#     # NUM_YTICKS = 13
+#     # LABELS_FONTSIZE = 12
+#     # LEGEND_FONTSIZE = 15
+#     # BAR_WIDTH = 0.85
+#     # FPOS_COLOR = "#b0b000"
+#     # FNEG_COLOR = "#e6194b"
 
-        #     # add ticks with class names to the conf matrix
-        # y_ticks = [nm + " (acc="+ self._float_to_str(acc) + ")"
-        #            for nm, acc in tickdata]
-        # x_ticks = [nm for nm, _ in tickdata]
-        # ax.set_yticks(range(conf_arr.shape[0]))
-        # ax.set_yticklabels(y_ticks, fontsize=self.TICKS_FONTSIZE,
-        #                    family="DejaVu Sans")
-        # ax.set_xticks(range(conf_arr.shape[1]))
-        # ax.set_xticklabels(x_ticks, fontsize=self.TICKS_FONTSIZE,
-        #                    family="DejaVu Sans", rotation=30, ha="right")
+
+#     def __init__(self, arr, mvn, max_datapoints=10000, samplerate=None):
+#         """
+#         """
+#         super().__init__(arr, max_datapoints, samplerate)
+#         self.mvn = mvn
+
+#     # def make_fig(self, fig_title, audio_arr, total, x_shared):
+#     #     """
+#     #     """
+#     #     # define and configure plt figure
+#     #     fig = plt.figure(figsize=self.FIG_ASPECT_RATIO)
+#     #     gs = list(gridspec.GridSpec(total, 1))
+#     #     axes = [plt.subplot(gs[0])]
+#     #     # add shared-x axes
+#     #     axes += [plt.subplot(gsi, sharex=axes[0])
+#     #              for gsi in gs[1: x_shared]]
+#     #     # add remaining axes:
+#     #     axes += [plt.subplot(gsi) for gsi in gs[x_shared:]]
+
+#     #     #
+#     #     info = "\nNo. axes=%d, dep. axes=%d" % (total, x_shared)
+#     #     fig.suptitle(fig_title + info, fontsize=self.FIG_TITLE_FONTSIZE,
+#     #                  fontweight="bold")
+#     #     fig.canvas.set_window_title(fig_title)
+
+#     #     axes[0].plot(audio_arr)
+
+#     #     for ax in axes[1:]:
+#     #         ax.plot(range(10), range(10))
+#     #     fig.subplots_adjust(**self.FIG_MARGINS)
+#     #     return fig
+
+
+#     def make_fig(self, num_xticks=10, num_yticks=10, tick_fontsize=7):
+#         """
+#         :returns: A matplotlib Figure containing the array given at
+#           construction. The interactive plot of the figure will react
+#           to the user's zooming by showing approximately
+#           ``self.max_datapoints`` number of samples.
+
+#         Reference:
+#         https://matplotlib.org/3.1.0/gallery/event_handling/resample.html
+#         """
+#         fig, ax = plt.subplots()
+#         self.axx, = ax.plot(self.arange, self.arr, '-')
+#         # set number of ticks
+#         ax.locator_params(axis="x", nbins=num_xticks)
+#         ax.locator_params(axis="y", nbins=num_yticks)
+#         #
+#         ax.xaxis.tick_top()
+#         plt.setp(ax.xaxis.get_majorticklabels(), rotation=30,
+#                  fontsize=tick_fontsize, family="DejaVu Sans", ha="left")
+#         plt.setp(ax.yaxis.get_majorticklabels(),
+#                  fontsize=tick_fontsize, family="DejaVu Sans")
+#         ax.set_autoscale_on(False)  # Otherwise, infinite loop
+#         ax.callbacks.connect('xlim_changed', self._update_ax)
+#         ax.set_xlim(0, len(self.arr))
+#         ax.xaxis.grid(True)
+#         if self.samplerate is not None:
+#             f = plt.FuncFormatter(lambda val, pos:
+#                                   sample_to_timestamp_formatter(
+#                                       val, pos, self.samplerate))
+#             ax.xaxis.set_major_formatter(f)
+#         #
+#         return fig